refactor(TextInput): migrate to TypeScript
Converts TextInput to TSX with a TextInputProps interface extending
native input attributes, drops PropTypes, and exports TextInputProps
from src/index.ts.

Notes on typing decisions (runtime unchanged):
- "date" added to the type union; DateInput already passes type="date"
- startIcon/endIcon typed as IconName, matching IconButton
- error typed string | string[] to match the previous propTypes
- showClearButton passed to Input coerced with Boolean() (was a
  truthy string before; Input only uses it as a conditional)
